added compress, decompress function, refactored code
