+++ src/utils/index.js
@@ -6,8 +6,8 @@ export { default as handleCd } from './handlers/cd.js';
 export { default as handleCat } from './handlers/cat.js';
 export { default as handleAdd } from './handlers/add.js';
 export { default as handleRn } from './handlers/rn.js';
-export { default as handleCp } from './handlers/cp.js';
-export { default as handleMv } from './handlers/mv.js';
+export { default as handleCpOrMv } from './handlers/cpOrMv.js';
 export { default as handleRm } from './handlers/rm.js';
 export { default as handleHash } from './handlers/hash.js';
+export { default as handleCompressOrDecompress } from './handlers/compressOrDecompress.js';
 export { default as getOsInfo } from './os_utils/os.js';
